requirements.txt
asgiref==3.2.7
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
confusable-homoglyphs==3.2.0
django-bootstrap4==1.1.1
django-paypal==1.0.0
django-registration==3.1
idna==2.9
Pillow==8.1.1
psycopg2==2.8.4
pytz==2018.7
requests==2.23.0
six==1.14.0
soupsieve==2.0
Django==3.1.6
pytz==2019.3
sqlparse==0.3.1
urllib3==1.25.8
